requirements.txt
asgiref==3.5.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
docutils==0.18
filter==0.0.0.20200724
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
keyring==23.2.1
kiwisolver==1.3.2
Markdown==3.3.6
matplot==0.1.9
matplotlib==3.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.21.3
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pkginfo==1.7.1
protobuf==3.20.0
Pygments==2.10.0
pyloco==0.0.139
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
scikit-learn==1.0.1
scipy==1.7.1
SimpleWebSocketServer==0.1.1
six==1.16.0
sqlparse==0.4.2
svgwrite==1.4.1
threadpoolctl==3.0.0
tqdm==4.62.3
Tree==0.2.4
twine==3.6.0
typing==3.7.4.3
tzdata==2022.1
urllib3==1.26.7
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.2.1
zipp==3.6.0
